drivers: Update uvbuilder for vs2019
Enable all warnings in uvbuilder.vcxproj by default with exceptions
listed in <DisableSpecificWarnings>. This caused warning-generated
errors that weren't there before, most of them from WDK headers. For
warnings out of our control, #pragma warning(disable:XXXX) are added to
silence them.

+++ deps/hypervisor/bfsdk/include/bfsupport.h
@@ -152,7 +152,7 @@ struct crt_info_t {
     uintptr_t arg2;
     uintptr_t arg3;
 
-    int info_num;
+    uint64_t info_num;
     struct section_info_t info[MAX_NUM_MODULES];
 
     uintptr_t func;
